+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Changelog
 ------------
 
+#### 6.0.7
+Added key handling to sortable `FlexTable` headers so that ENTER and SPACE keys can be used to toggle sort direction.
+
+#### 6.0.6
+Added conditional checks to when `aria-label`, `role`, and `tabIndex` get attached to `FlexTable` headers and rows.
+These a11y properties are only added when on-click or sort handlers are present.
+
+#### 6.0.5
+Added `aria-label` and `role` attributes to `FlexTable`, `Grid`, and `VirtualScroll` components to fix a11y issues reported by [reactjs/react-a11y](https://github.com/reactjs/react-a11y).
+Thanks to @globexdesigns for the contributions!
+
+#### 6.0.4
+Separated horiontal and vertical `Grid` metadata calculation to avoid unnecessarily recomputing row metadata for `FlexTable`s and `VirtualScroll`s when a browser's window is resized, for example.
+Also replaced `columnWidth` and `rowHeight` getter uses in `Grid.render` in favor of cached cell metadata instead.
+
 ##### 6.0.3
 Small update to `FlexTable` to move the `rowGetter` call outside of the column loop to reduce the number of times that method gets called.